fix TunEndpoint setup (to make DNS work again)
In a previous commit, TunEndpoint::configure was folded into its
constructor.  This would be fine, but at the end of that was
TunEndpoint::setup_dns, which passes a weak_ptr of itself to another
class.  In the future *that* should maybe change, but for now setup_dns
will be called *after* a shared_ptr to TunEndpoint is constructed.