Add LOKINET_LIBRARY_ONLY CMake option
This option conditionally doesn't compile a majority of code that is
only used for full/tun clients and relays.  There is more that can still
be removed (e.g. anything behind "is service node"), but this should be
sufficient that lokinet can compile library-only on any platform without
issue.

Some of the changes are a bit clumsy and should be revisited later, but
for the most part it's as clean as possible.