Accept RouterID if majority have it
This will need some tuning, but for testing I went for the simple
implementation.  The idea here is that instead of comparing entire sets
of RouterIDs and taking the most common set (if no tie), we should keep
every RouterID where a sufficient number of responses agree that we
should have it.  In this commit, "sufficient" is more than half of
responses.