Don't allow IP range restrictions to prevent edge selection
If we are running with pinned edges or very few edges then it is
possible that a pivot we need to connect to coincides with the pivot.
In that case, we need to violate the IP range exclusion for edge
selection to avoid having no possible path at all (but still keep it for
all other hops).