Edge conns: make in-progress connections count
When we do the checks for "Client connecting to 4 random routers to keep
alive (current:0, needed:4)" we weren't counting inactive (i.e.
in-progress) connections, and so unless all the connections established
within a single router tick we would over-connect to edges.

This fixes it to not ignore in-progress edge connections.