Don't use std::ranges filter for first hop selection
We want the filter conditions here to be called only once per edge (both
as a slight optimization, but also so that the log statements make
sense), but a filter doesn't guarantee that (nor does it do it in
practice).

This replaces it with a simpler for loop that calls each value exactly
once and uses simple single-element reservoir sampling to choose the
element.