Drop PathBuildLimiter
The limiter was a stable lokinet holdover to try to address Lokinet path
issues by forcing them to slow down, but just gets in the way in new
lokinet.  This removes the limiter.

Since this was only ever client-side, the limiter doesn't protect
against intentional abusive behaviour, and a client that rebuilds lots
of paths really isn't an issue without old Lokinet's wire protocol.