Drop Path.intro; re-add path/hop expiry to path builds
The ClientIntro intro inside Path was confusing: sometimes its expiry
stored the path expiry, but sometimes the actual client intro expiry.
Moreover it isn't needed: ClientIntros are a nearly trivial type, and so
constructing one on the fly when needed is much easier.  Thus this
commit drops it, replacing it with a `make_intro` that constructs an
intro on the fly.

This also puts re-adds path and hop expiries as a path build field so
that clients can signal to the paths they build through when a path
lifetime is less than the maximum.  (The field is optional; omitting it
means max lifetime).

Other misc. changes/fixes:
- Move TransitHop decryption and serialization into PathHandler, so that
  it is alongside the PathHandler code that builds the transithop
  frames.
- Set the path expiry at Path construction, rather than when the path is
  established.  If there's some network lag in establishing a path, we
  want to use the more conservative expiry.