Distributed inbound path lifetimes; fix publishing
This commit distributes the lifetime of built inbound paths to avoid the
possibility of a cluster of paths that could expire at the same time,
thus resulting in a failed connection.

This basically tries to maintain paths with 5,10,15,20 expiries at all
times; if a path dies or expires and a new path is needed, we set the
path lifetime at whatever it needs to be to fit in with the above
pattern, essentially filling the gap.

See the added comments in handlers/session.cpp about how it all works.

To accomplish that, this adds a path lifetime field to the path build so
that relays don't end up keeping paths around longer than the client
that created the path.

This also reworks how CC publishing works: rather than using a fixed
timer, we publish whenever we complete a final path that brings us up to
our paths target, so that we publish immediately upon path changes.

Also reworked path ping logs to show more ping stats (in debug output).