Untangle failing connection spaghetti
Relay-to-relay connections spuriously fail in the first few minutes
because of mishandling of the stream and connection in the case of
racing connections.

Fundamentally this was happening because we would open a stream
immediately upon establishing a connection, store the stream, and try
sending down the stream but there are multiple problems with that:
- the connection might not get established at all (because of the
  attempt to avoid simultaneous connections in the key verify callback).
- subsequent established connections weren't always handled properly.
- recent changes that removed random delays before establish connections
  broke things because that apparently was a kludge to pretend the bugs
  didn't exist.

This rewrite the simultaneous connection logic entirely so that:

- parallel (1 inbound, 1 outbound) connections *can* exist at the same
  time.  When this state happens, we note the time where we started
  having two live connections, and always set the stream to the
  associated stream of the winning connection.
- We periodically check for any such parallel connections and close the
  non-winner, but only after (at least) 15s has passed, so that any
  stream data sent down the loser connection before both were establish
  still has a change to replied to.

As part of this effort, this commit splits link/link_manager.cpp
(containing llarp::link::Manager, renamed from llarp::LinkManager) into
two: link/endpoint.cpp, and link/link_manager.cpp, where link::Endpoint
is now in charge of everything related to the actual endpoint, and
Manager is now only in charge of higher level details (such as path and
session messageS).  Previously there was a very unclear split between
the two classes.

Additionally, service node testing seemed entirely broken: it would
establish and close connections, but that almost surely breaks data
travelling between those nodes when the estalish happens *or* the close
happens.  This disables it for now with a TODO FIXME critical log
statement.

Bootstrapping was also broken by this commit, but seems to need to be
replaced (and so in this commit is simply commented out, to be replaced
in the following commit).

Other minor changes:
- clients were erroneously listening for incoming connections.
- link_manager.cpp was compiled twice (once in the `lokinet-core` lib,
  once in `lokinet-wire`).  Having it only on lokinet-wire doesn't work
  (it relies on things in core), so this drops lokinet-wire and moves
  the link/* into lokinet-core.
- replace nlohmann/json.hpp with nlohmann/json_fwd.hpp in several
  top-level headers (relay_contact.hpp, router_id.hpp), from which it
  was ending in *must* compilation units.
- Made intermediate Session class (OutboundSession, InboundSession)
  destructors protected so that the classes can't be directly
  instantiated.