Add global log cat; improve some logging; unify buildstats
This adds a `log_global` log category, and sets it to info level (even
if the generic level is set higher) so that it can be used for messages
we want to show that really aren't warnings or errors.

One such new log statement is an global info (or warning) when
connectivity is gained (or lost).  This required making client-connected
a state variable of link endpoint, which is now used instead of a silly
not-quite-right bool in router.cpp.

Tweak startup logs to show various details, including the the lokinet
address & local private range in a global info message during startup.

Another is to change the current stats line to make it a global info
(instead of critical), and also to make it more useful info:
    - it now shows inbound/outbound(client)/outbound(relay) session
      counts
    - it now shows inbound and outbound path counts properly
    - it now shows total path build attempts and success rate
    - report every 1min rather than an excessive 10s.

Along the way, I noticed that the path::BuildStats was unique per
PathBuilder (which means each outbound session got its own counter).
This makes no sense, so now it is global and lives in Router.