Fix inbound stream setup to happen in network loop
We were queuing the incoming stream too late by doing it inside the
router loop via a `call` from the network loop, but that is too late
because early stream data can easily arrive before the router loop call
runs which ends up creating a default stream for stream 0, and *then*
the queued stream is set up for the next stream (which is never used).

As a result of this, the server has a race condition which, if it
happens as described above, results in all BTRequestStream commands send
to the node failing because there is a BTReqStream on the other end.