Fix inbound relay-relay stream delays
Under QUIC, streams do not exist on the stream-receiver end of a
connection until stream data is actually received.  This meant that
relay-to-relay connections could get stuck with all data sent back on
that stream just getting queued into the stream because the stream
itself is stuck as a pending incoming stream object.

This fixes it by immediately queuing a tiny "noop" non-responding,
do-nothing command when we open an outbound connection stream to force
it open at the other end so that the other end can talk to us before we
have anything to say to the other end.

Without this, stream data could get delayed for minutes, and the network
remained unreliable for the first 10 minutes or so after an all-node
testnet restart (because it takes that long for node gossipping to kick
in, which could be the initial data that force opens the streams).