Initiate RC lookup in OutboundRelaySession constructor
This parallels the CC lookup behaviour of OutboundClientSession, and is
a necessary step towards removing the lookups from tun.cpp.

It also allows us to tick the path to start path building immediately if
we already have the RC (or upon receiving it, if we don't) rather than
delaying up to 250ms for the next tick to start building.