Bootstrap: return all RCs; no push RC; add zstd compression
Drop RemoteRC from relay bootstrap requests.  It appears to have been
added thinking it was necessary to get the network off the ground, but
that's only because of a bug with nodedb not properly including its own
RC.  That bug should be fixed instead, so that the bootstrapping node
will at least provide an updated RC for the bootstrap itself, and then a
regular RC broadcast will go to the bootstrap, and from there to
everyone else, who can then connect inwards and let the bootstrapping
node accumulate all network RCs within an RC broadcast cycle.

This commit also drops the "quantity" request parameter from bfetch_rcs:
when we need to bootstrap, we realistically need to get everything so
that we can actually make things like strict-connect and IP selection
work, and so bfetch_rcs now returns all RCs.

Finally this adds zstd as a dependency (which, for embedded libsession,
will already be present), and uses it to compress RCs returned from a
bootstrap fetch.  The pubkey, signature, and IP (100 bytes total) aren't
very compressible, but the remainder of the 155 byte typical RC is full
of repeated values and bt-encoding bytes and compresses to almost
nothing, resulting in about a 1/3 reduction in size in practice, even at
very low compression levels.

Also removes LinkManager's unused fetch_bootstrap_rcs's method.