Rename RPCClient -> OxendRPC; upgrade to oxenmq 1.3
RPCClient was a confusing name, because it isn't in any way an actual
lokinet RPC client, but rather is a highly special case class for
talking to oxend.

OxenMQ 1.3 finally adopts oxen-logging, so this updates for that
compatibility as well, which means dropping the now-gone logging
callbacks in oxenmq constructors.