Add synchronous sn list fetch (w/ fallback) on relay startup
When relays start, they sometimes get incoming connections before they
have fetched the registered SN list, and as a result reject them.

This adds a synchronous wait of up for 10s for the initial fetch
*before* we start up the quic machinery to avoid this.

If we don't get an update within 10s, this also adds a fallback that
loads router IDs from all known RCs in the nodedb so that lokinet can
still start up even if oxend is temporarily unavailable.