Re-implement router testing
This reimplements router testing for dev lokinet.

We now test one random node every 1/3 of a second, and retest up to four
currently-failing nodes on that same interval.

The testing is done by establishing a *client* connection (i.e. using
the client ALPN) to the relay being tested, issuing a (new) "ping"
command on it, and awaiting the response (which must, including
connection time, come back within 4s).