Take edge range into account when selecting inbound path terminus
If all edges share the same unique ip range -- perhaps because we only
have one edge, or pinned edges all in the same range, or just by luck --
then we can end up in a situation where pure random selection could
choose a terminus that is also in that same range, which ends up
violating the unique path requirement.

The violation happens because the path builder is told the terminus to
build a path to, and is also restricted in first edge selection to
established edges, and so cannot avoid a unique range violation when all
edges and terminus are in the same range.

This adds detection of that case so that we can avoid selecting such a
terminus.