Smarter and fixed first hop selection
This fixes some problems in first hop selection:
- Path unique IP range restrictions were not being applied.
- In 2-hop path mode, we could end up selecting the same terminus and
  edge, but that is an invalid path.

This addresses these by switching to a three-attempt selection process
for edge selection: first with range restrictions applied to try to
avoid the same network as the terminus; then falling back to allowing a
range violation (but not the pivot itself); then falling back to using
the pivot itself as the edge.

If we end up needing such a fallback, we now extend the path length by 1
to compensate for the violation, so that you still end up with a path
with the same number of *distinct* network hops on it.

This also ends up fixing 2-hop mode by forcing it to become a 3-hop mode
if the only available edge is also the terminus, so that we avoid the
invalid client-A-A path and instead force a valid client-A-B-A path.