Fix rare off-by-one in inbound path build logic
There was an incredibly rare case where SessionEndpoint::update_paths
could get past the "we need more paths" check because of a timing issue
(not passing "now" to num_paths()), and then the code would index an
array at -1, which would corrupt the stack in a way that led to some
crashing.  This is fixed by passing "now" correctly to that call and
correcting a conditional later which was simply wrong/useless anyway.

In addition, the other reason this corruption was possible is that the
code was trying to not build a path in the first "time bucket" (e.g. 0-5
min expiration), but this was implemented in a way which has a few
undesirable side-effects.  Instead, we now *do* build a path in that
"bucket" if it doesn't have one and should.