path switch backup session init
In the event that the remote end of an outbound session restarts, it
would be good to be able to "resume" that session quickly and
transparently.  As such, path switch messages will now also include a
session init message so that if the remote does not have a session
for the path switch, it will make a new session and the initiator can
carry on as normal.

TODO: The outbound session does not currently do anything
special/different with the "session accept" message it will receive in
response to this session init.