Fix the slot0 fixed fix again
Something was still wrong with the slot0 fix in that slot0 was
referencing the wrong slot (i.e. was a slot start point 0-5min ago), and
as a result we ended up with paths in the 15-20m slot being considered
too large relative to slot0 and producing warnings.

The problem was that slot0 was wrong, and was actually referencing
slot-minus-1, and our first slot for consideration (i.e. slot 0) needs
the +1 to get us to the first non-negative slot position so that slots 0
through 3 map to times in the [0-5], [5-10], [10-15], [15-20] minute
ranges.

Also rewrites the comment to better explain the +1 because the original
explanation I wrote confused us to the point of a first fix (and update
to that fix) that broke things.