Optionally allow inbound pivot reuse
Currently we have a hard cap of inbound paths at the number of known
RCs; even with the "extra" option to go beyond 10, once you have a path
to every pivot, no more inbound paths can be built.

This adds an inbound path option allowing you to reuse pivots up to N
times (where N=1 remains the default), so that in special cases you can
maintain more than one path to a pivot.

This also switches pivot selection to use a multi-pass pivot selection
to make that work, particularly on testnet where the target number of
paths can easily exceed the number of RCs: get_n_random_rcs() can only
return at most N (= number of known RCs), but if we have pivot reuse we
might actually want more than that.

This puts pivot selection into a loop so that if we don't get enough
pivots to reach our target, we try again until either we get enough or
we get none (in which case we can warn about it).